pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "streamvis"
version = "0.1.0"
description = "Interactive visualizations of streaming data using Bokeh"
authors = [{name="Henry Bigelow", email="hrbigelow@gmail.com"}]
requires-python = ">=3.12.0"
dependencies = [
  "tornado",
  "numpy",
  "bokeh>=3.0.0",
  "google-cloud-storage",
  "grpcio>=1.71.0",
  "grpcio-tools>=1.71.0",
  "hydra-core>=1.3.2",
]

[project.scripts]
streamvis = "streamvis.script:main"

[tool.setuptools]
packages = ["streamvis"]



